eth_sendTx tests
