+++ packages/eth-rpc-adapter/src/__tests__/endpoint-tests/endpoint.ts
@@ -616,399 +616,6 @@ describe('endpoint', () => {
     });
   });
 
-  describe('eth_sendRawTransaction', async () => {
-    const chainId = BigNumber.from((await eth_chainId()).data.result).toNumber();
-
-    const endpoint = process.env.ENDPOINT_URL || 'ws://127.0.0.1:9944';
-    const wsProvider = new WsProvider(endpoint);
-    const api = await ApiPromise.create({ provider: wsProvider });
-
-    const genesisHash = api.genesisHash.toHex(); // TODO: why EIP-712 salt has to be genesis hash?
-
-    const ETH_Digits = 18;
-    const ACA_Digits = 12;
-    const TX_FEE_OFF_TOLERANCE = parseEther('0.01').toBigInt(); // 0.01 ACA
-
-    const queryEthBalance = async (addr): Promise<BigNumber> =>
-      BigNumber.from((await eth_getBalance([addr, 'latest'])).data.result);
-
-    const queryNativeBalance = async (addr: string) =>
-      (await queryEthBalance(addr)).div(10 ** (ETH_Digits - ACA_Digits));
-
-    const getTxFeeFromReceipt = async (txHash: string, toNative = false): Promise<bigint> => {
-      await sleep(789); // give cache/subquery a little bit time
-      const { gasUsed, effectiveGasPrice } = (await eth_getTransactionReceipt([txHash])).data.result;
-
-      const calculatedTxFee = BigInt(gasUsed) * BigInt(effectiveGasPrice);
-
-      return toNative
-        ? calculatedTxFee / BigInt(10 ** (ETH_Digits - ACA_Digits))
-        : calculatedTxFee;
-    };
-
-    afterAll(async () => {
-      await api.disconnect();
-    });
-
-    describe('deploy contract (hello world)', () => {
-      const partialDeployTx = {
-        chainId,
-        data: deployHelloWorldData,
-        type: 0,
-      };
-
-      describe('with legacy EIP-155 signature', () => {
-        it('send tx correctly, gas estimation is accurate, receipt\'s gas info is accurate', async () => {
-          const prevBalance = await queryEthBalance(wallet1.address);
-
-          const unsignedTx = {
-            ...partialDeployTx,
-            nonce: await getNonce(wallet1.address),
-          };
-
-          const { gasPrice, gasLimit } = await estimateGas(unsignedTx);
-
-          const rawTx = await wallet1.signTransaction({
-            ...unsignedTx,
-            gasPrice,
-            gasLimit,
-          });
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.equal(undefined); // for TX error RPC will still return 200
-
-          const receiptTxFee = await getTxFeeFromReceipt(res.data.result);
-          const afterBalance = await queryEthBalance(wallet1.address);
-
-          const realTxFee = prevBalance.sub(afterBalance).toBigInt();
-          const estimatedTxFee = gasPrice.mul(gasLimit).toBigInt();
-          const diffReceiptTxFee = bigIntDiff(realTxFee, receiptTxFee);
-          const diffEstimateTxFee = bigIntDiff(realTxFee, estimatedTxFee);
-
-          // console.log({
-          //   estimatedTxFee: formatEther(estimatedTxFee),
-          //   realTxFee: formatEther(realTxFee),
-          //   receiptTxFee: formatEther(receiptTxFee),
-          //   diffReceiptTxFee: formatEther(diffReceiptTxFee),
-          //   diffEstimateTxFee: formatEther(diffEstimateTxFee),
-          // });
-
-          expect(diffReceiptTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-
-          // estimated tx fee is slightly overestimated now
-          expect(diffEstimateTxFee < TX_FEE_OFF_TOLERANCE).to.be.false;
-        });
-      });
-
-      describe('with EIP-1559 signature', () => {
-        it('throw correct error', async () => {
-          const unsignedTx = {
-            ...partialDeployTx,
-            nonce: await getNonce(wallet1.address),
-            gasPrice: undefined,
-            maxPriorityFeePerGas: BigNumber.from(0),
-            maxFeePerGas: ONE_HUNDRED_GWEI,
-            type: 2,
-          };
-
-          const rawTx = await wallet1.signTransaction(unsignedTx);
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.contain('unsupported transaction type: 2, please use legacy or EIP-712 instead');
-        });
-      });
-
-      describe('with EIP-712 signature', () => {
-        it('send tx correctly, receipt\'s gas info is accurate', async () => {
-          const prevBalance = await queryEthBalance(wallet1.address);
-
-          const gasLimit = BigNumber.from('210000');
-          const validUntil = 10000;
-          const storageLimit = 100000;
-
-          const unsignEip712Tx = {
-            ...partialDeployTx,
-            nonce: await getNonce(wallet1.address),
-            salt: genesisHash,
-            gasLimit,
-            validUntil,
-            storageLimit,
-            type: 0x60,
-          };
-
-          const sig = signTransaction(account1.privateKey, unsignEip712Tx as AcalaEvmTXPayload);
-          const rawTx = serializeTransaction(unsignEip712Tx as UnsignedAcalaEvmTX, sig);
-          const parsedTx = parseTransaction(rawTx);
-
-          expect(parsedTx.gasLimit.eq(gasLimit)).equal(true);
-          expect(BigNumber.from(parsedTx.validUntil).eq(validUntil)).equal(true);
-          expect(BigNumber.from(parsedTx.storageLimit).eq(storageLimit)).equal(true);
-
-          expect(parsedTx.from).equal(wallet1.address);
-          expect(parsedTx.data).equal(deployHelloWorldData);
-          expect(parsedTx.type).equal(96);
-          expect(parsedTx.maxPriorityFeePerGas).equal(undefined);
-          expect(parsedTx.maxFeePerGas).equal(undefined);
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.equal(undefined); // for TX error RPC will still return 200
-
-          const receiptTxFee = await getTxFeeFromReceipt(res.data.result);
-          const afterBalance = await queryEthBalance(wallet1.address);
-
-          const realTxFee = prevBalance.sub(afterBalance).toBigInt();
-          const diffReceiptTxFee = bigIntDiff(realTxFee, receiptTxFee);
-
-          // console.log({
-          //   realTxFee: formatEther(realTxFee),
-          //   receiptTxFee: formatEther(receiptTxFee),
-          //   diffReceiptTxFee: formatEther(diffReceiptTxFee),
-          // });
-
-          expect(diffReceiptTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-        });
-      });
-    });
-
-    describe('call contract (transfer ACA)', () => {
-      const iface = new Interface(TokenABI.abi);
-      const transferAmount = parseUnits('123.321', ACA_Digits);
-      const partialTransferTX = {
-        chainId,
-        from: wallet1.address,
-        to: ADDRESS.ACA,
-        data: iface.encodeFunctionData('transfer', [account2.evmAddress, transferAmount]),
-        type: 0,
-      };
-
-      describe('with legacy EIP-155 signature', () => {
-        it('has correct balance after transfer, and receipt\'s gas info is accurate', async () => {
-          const [balance1, balance2] = await Promise.all([
-            queryNativeBalance(account1.evmAddress),
-            queryNativeBalance(account2.evmAddress),
-          ]);
-
-          const transferTX = {
-            ...partialTransferTX,
-            nonce: await getNonce(wallet1.address),
-          };
-
-          const { gasPrice, gasLimit } = await estimateGas(transferTX);
-          const rawTx = await wallet1.signTransaction({
-            ...transferTX,
-            gasPrice,
-            gasLimit,
-          });
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.equal(undefined); // for TX error RPC will still return 200
-
-          await sleep(1000);
-          const [_balance1, _balance2] = await Promise.all([
-            queryNativeBalance(account1.evmAddress),
-            queryNativeBalance(account2.evmAddress),
-          ]);
-
-          const receiptTxFee = await getTxFeeFromReceipt(res.data.result);
-          const realTxFee = nativeToEthDecimal(balance1.sub(_balance1).sub(transferAmount).toBigInt()).toBigInt();
-          const estimatedTxFee = gasPrice.mul(gasLimit).toBigInt();
-          const diffReceiptTxFee = bigIntDiff(realTxFee, receiptTxFee);
-          const diffEstimateTxFee = bigIntDiff(realTxFee, estimatedTxFee);
-
-          // console.log({
-          //   estimatedTxFee: formatEther(estimatedTxFee),
-          //   realTxFee: formatEther(realTxFee),
-          //   receiptTxFee: formatEther(receiptTxFee),
-          //   diffReceiptTxFee: formatEther(diffReceiptTxFee),
-          //   diffEstimateTxFee: formatEther(diffEstimateTxFee),
-          // });
-
-          expect(diffReceiptTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-          expect(diffEstimateTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-          expect(formatEther(_balance2.sub(balance2))).to.eq(formatEther(transferAmount));
-        });
-      });
-
-      describe('with EIP-1559 signature', () => {
-        it('throw correct error', async () => {
-          const priorityFee = BigNumber.from(0);
-          const transferTX = {
-            ...partialTransferTX,
-            nonce: await getNonce(wallet1.address),
-            gasPrice: undefined,
-            maxPriorityFeePerGas: priorityFee,
-            maxFeePerGas: 1000000,
-            type: 2,
-          };
-
-          const rawTx = await wallet1.signTransaction(transferTX);
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.contain('unsupported transaction type: 2, please use legacy or EIP-712 instead');
-        });
-      });
-
-      describe('with EIP-712 signature', () => {
-        it('has correct balance after transfer, and receipt\'s gas info is accurate', async () => {
-          const [balance1, balance2] = await Promise.all([
-            queryEthBalance(account1.evmAddress),
-            queryEthBalance(account2.evmAddress),
-          ]);
-
-          const gasLimit = BigNumber.from('210000');
-          const validUntil = 10000;
-          const storageLimit = 100000;
-
-          const transferTX = {
-            ...partialTransferTX,
-            nonce: await getNonce(wallet1.address),
-            salt: genesisHash,
-            gasLimit,
-            validUntil,
-            storageLimit,
-            type: 0x60,
-          };
-
-          const sig = signTransaction(account1.privateKey, transferTX as AcalaEvmTXPayload);
-          const rawTx = serializeTransaction(transferTX as UnsignedAcalaEvmTX, sig);
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.equal(undefined); // for TX error RPC will still return 200
-
-          await sleep(1000);
-          const [_balance1, _balance2] = await Promise.all([
-            queryEthBalance(account1.evmAddress),
-            queryEthBalance(account2.evmAddress),
-          ]);
-
-          const receiptTxFee = await getTxFeeFromReceipt(res.data.result);
-          const realTxFee = balance1.sub(_balance1).sub(nativeToEthDecimal(transferAmount)).toBigInt();
-          const diffReceiptTxFee = bigIntDiff(realTxFee, receiptTxFee);
-
-          // console.log({
-          //   realTxFee: formatEther(realTxFee),
-          //   receiptTxFee: formatEther(receiptTxFee),
-          //   diffReceiptTxFee: formatEther(diffReceiptTxFee),
-          // });
-
-          expect(diffReceiptTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-          expect(formatEther(_balance2.sub(balance2))).to.eq(formatEther(nativeToEthDecimal(transferAmount)));
-        });
-      });
-    });
-
-    describe('send native ACA token', () => {
-      const transferAmount = parseEther('16.88');
-      const partialNativeTransferTX = {
-        chainId,
-        from: wallet1.address,
-        to: account2.evmAddress,
-        value: transferAmount,
-      };
-
-      describe('with legacy EIP-155 signature', () => {
-        it('has correct balance after transfer, and receipt\'s gas info is accurate', async () => {
-          const [balance1, balance2] = await Promise.all([
-            queryEthBalance(account1.evmAddress),
-            queryEthBalance(account2.evmAddress),
-          ]);
-
-          const unsignedTx = {
-            ...partialNativeTransferTX,
-            nonce: await getNonce(wallet1.address),
-          };
-          const { gasPrice, gasLimit } = await estimateGas(unsignedTx);
-
-          const rawTx = await wallet1.signTransaction({
-            ...unsignedTx,
-            gasPrice,
-            gasLimit,
-          });
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.equal(undefined); // for TX error RPC will still return 200
-
-          await sleep(1000);
-          const [_balance1, _balance2] = await Promise.all([
-            queryEthBalance(account1.evmAddress),
-            queryEthBalance(account2.evmAddress),
-          ]);
-
-          const receiptTxFee = await getTxFeeFromReceipt(res.data.result);
-          const realTxFee = balance1.sub(_balance1).sub(transferAmount).toBigInt();
-          const estimatedTxFee = gasPrice.mul(gasLimit).toBigInt();
-          const diffReceiptTxFee = bigIntDiff(realTxFee, receiptTxFee);
-          const diffEstimateTxFee = bigIntDiff(realTxFee, estimatedTxFee);
-
-          // console.log({
-          //   estimatedTxFee: formatEther(estimatedTxFee),
-          //   realTxFee: formatEther(realTxFee),
-          //   receiptTxFee: formatEther(receiptTxFee),
-          //   diffReceiptTxFee: formatEther(diffReceiptTxFee),
-          //   diffEstimateTxFee: formatEther(diffEstimateTxFee),
-          // });
-
-          expect(diffReceiptTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-          expect(diffEstimateTxFee < TX_FEE_OFF_TOLERANCE).to.be.true;
-          expect(formatEther(_balance2.sub(balance2))).to.eq(formatEther(transferAmount));
-        });
-      });
-
-      describe('with EIP-1559 signature', () => {
-        it('throw correct error', async () => {
-          const transferTX = {
-            ...partialNativeTransferTX,
-            nonce: await getNonce(wallet1.address),
-            gasPrice: undefined,
-            maxPriorityFeePerGas: 0,
-            maxFeePerGas: 10000000000,
-            type: 2,
-          };
-
-          const rawTx = await wallet1.signTransaction(transferTX);
-
-          const res = await eth_sendRawTransaction([rawTx]);
-          expect(res.data.error?.message).to.contain('unsupported transaction type: 2, please use legacy or EIP-712 instead');
-        });
-      });
-
-      describe('with EIP-712 signature', () => {
-        // TODO: EIP-712 doesn't use ETH gasLimit and gasPrice, do we need to support it?
-        it.skip('has correct balance after transfer', async () => {
-          // const balance1 = await queryEthBalance(account1.evmAddress);
-          // const balance2 = await queryEthBalance(account2.evmAddress);
-          // const gasLimit = BigNumber.from('210000');
-          // const validUntil = 10000;
-          // const storageLimit = 100000;
-          // const transferTX: AcalaEvmTX = {
-          //   ...partialNativeTransferTX,
-          //   ...(await estimateGas()),
-          //   nonce: await getNonce(wallet1.address),
-          //   salt: genesisHash,
-          //   gasLimit,
-          //   validUntil,
-          //   storageLimit,
-          //   type: 0x60
-          // };
-          // const sig = signTransaction(account1.privateKey, transferTX);
-          // const rawTx = serializeTransaction(transferTX, sig);
-          // const parsedTx = parseTransaction(rawTx);
-          // const res = await eth_sendRawTransaction([rawTx]);
-          // expect(res.data.error?.message).to.equal(undefined); // for TX error RPC will still return 200
-          // const txHash = res.data.result;
-          // const { gasUsed, effectiveGasPrice } = (await eth_getTransactionReceipt([txHash])).data.result;
-          // const calculatedTxFee = BigInt(gasUsed) * BigInt(effectiveGasPrice) / BigInt(10 ** (ETH_Digits - ACA_Digits));
-          // const _balance1 = await queryEthBalance(account1.evmAddress);
-          // const _balance2 = await queryEthBalance(account2.evmAddress);
-          // const realTxFee = balance1.sub(_balance1).sub(transferAmount).toBigInt();
-          // const diff = bigIntDiff(realTxFee, receiptTxFee);
-          // expect(_balance2.sub(balance2).toBigInt()).equal(transferAmount.toBigInt());
-          // expect(Number(diff)).to.lessThan(TX_FEE_OFF_TOLERANCE);
-        });
-      });
-    });
-  });
-
   describe('eth_call', () => {
     const callRequest = (abi: any) => async (address: string, method: string, params?: any[], blockTag?: any) => {
       const iface = new Interface(abi);
