+++ packages/eth-rpc-adapter/src/__tests__/utils/index.ts
@@ -1,7 +1,8 @@
 export * from './subs-manager';
-export * from './test-utils';
+export * from './misc';
 export * from './evm-accounts';
 export * from './consts';
 export * from './gas-monster';
 export * from './eth-rpc-apis';
 export * from './erc20';
+export * from './setup';
